Quiz application
Create a native application for iPhone or iPad that tests the user
knowledge of a certain topic.
Your application should have at least 10 questions hard-corded  and
each time the user starts the application , 5 of these questions are
selected randomly.
For every question , user should select one option out of four options
presented.
When the user answers the fifth question, the application should
display the user’s score and one of the following message depending on
their score

1. Score 0/5,1/5 or 2/5 Message - > “ Please try again “
2. Score 3/5-> “good jog”
3. Score 4/5-> “Excellent work”
4. Score 5/5-> “you are a genius”

The topic for the questions should be closed carefully and all the
questions should be meaningful. All the questions should be based on
the topic you have chosen.
If the score of the test is Option1 user should be given an option to
retake the quiz.